fix(arenabench): unbreak main — split the transcript reader out of telemetry.py
My own #2391 did this. `telemetry.py` went 1474 -> 1534 lines and crossed the
1500-line gate, so `file size ratchet` now fails on `main` and therefore on
every PR in flight. It merged anyway, which is its own finding: the guard is
advisory at the merge button and only the author's push is watched.

AGENTS.md leaves no choice about the fix, and it is the right one:

  "A new file that crosses 1500 lines fails the gate outright, and the
   baseline accepts no new entries — a file approaching the limit gets split,
   not grown over it."

So: split, no baseline entry.

The seam is the file's own `# Transcripts` banner, not a line count. Reducing
trials to scoreboard numbers and replaying a trial's narration are different
jobs read by different consumers, and only two names cross between them —
which is why `_proof_line` moves too, being called from exactly one place,
inside the moved block.

  telemetry.py  1534 -> 1129
  transcript.py         429  (new)

`TranscriptReader` and `TranscriptState` are re-exported from `telemetry`
rather than repointed at their call sites: they are part of that module's
published surface (`server.py` and the test suite both import them from
there), and moving a file should not be an API break for readers who never
cared where it lived. Both import paths resolve to the same object, asserted
directly.

Pure code motion — no behaviour touched. The full arenabench suite passes
unchanged, `scripts/check-file-size.sh` reports OK, and ruff is clean on both
files (`field` was left behind in telemetry's imports and `distill` was never
used by the moved code; both removed by hand rather than by `--fix`, after
checking each was genuinely unreferenced).
